Fix bug when creating private overlays after loading an overlay.
The code currently copies the 'overlay' descriptor file from the overlay
directory to the user's data directory. If the user has previously saved a
private overlay, then an 'overlay' subdirectory will exist. In this case, perl
copy() will copy the overlay file to 'overlay/overlay'.

However if the user previously has not saved a private overlay, that
subdirectory won't exist, and so an 'overlay' entry in the user data directory
will be created, which is a file. At which point, when the user attempts to
save a private overlay under 'overlay/<id>', the operation will fail since
'overlay' is a file.

This change simply avoids copying the overlay descriptor file when loading an
overlay.